Move common types to common package
